Broke up query.ts into lexer and parser components
